refactor: extract ime-input.command
